+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
-## 0.1.15 - 2018-07.31
+## 0.1.16 - 2018-10-29
+### Changed
+- Added Spec support for all options, and for the subcommand at once.
+- Added keywords as a parse type (tks jwhitlark)
+
+
+## 0.1.15 - 2018-07-31
 ### Changed
 - Added EDN support (tks jwhitlark)
 - Added optional YAML parsing, but using it breaks Graal (tks jwhitlark)